requirements.txt
google-generativeai
google-cloud-aiplatform
streamlit
python-dotenv
pandas
yfinance
duckduckgo-search
langchain
langgraph
google-adk
